Added check for purge script

+++ scripts/purge_dist.sh
@@ -8,4 +8,13 @@ else
     echo Already at root folder.
 fi
 
+
+if [ ! -d dist ]
+then
+    echo The dist folder does not exist.
+    exit 1
+fi
+
 rm -r dist
+
+echo Removed dist successfully.
